backend/staff-service/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
PyJWT==2.9.0
requests==2.32.3
mysqlclient==2.2.4
psycopg2-binary==2.9.9
gunicorn==23.0.0
